get_motion_files: take channel as argument
For "TrackMix" with two lenses, channel 0 is the wide-angle lens while channel 1 is the telephotolens.

The "action" parameter of the Search request is documented in Reolink's PDF as being "0", yet this code passes "1". Not modified in this change as it did not seem to prevent correct operation on my trackmix wifi.